refactor: organize share-related files into a dedicated package
