+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/profile/ProfileScreen.kt
@@ -45,8 +45,8 @@ import net.thechance.mena.identity.presentation.screen.profile.components.OtherS
 import net.thechance.mena.identity.presentation.screen.profile.components.ProfileInfoContainer
 import net.thechance.mena.identity.presentation.screen.profile.components.ShareIcon
 import net.thechance.mena.identity.presentation.screen.profile.components.ThemeDialog
-import net.thechance.mena.identity.presentation.screen.profile.components.dialog.share.ShareQrCode
-import net.thechance.mena.identity.presentation.screen.profile.components.dialog.share.ShareSheet
+import net.thechance.mena.identity.presentation.screen.profile.components.share.ShareQrCode
+import net.thechance.mena.identity.presentation.screen.profile.components.share.utils.ShareSheet
 import org.jetbrains.compose.resources.stringResource
 
 class ProfileScreen : BaseScreen<
