+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/profile/components/share/ShareQrCodeUIEffect.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.presentation.screen.profile.components.dialog
+package net.thechance.mena.identity.presentation.screen.profile.components.share
 
 import org.jetbrains.compose.resources.StringResource
