+++ identity_presentation/src/iosMain/kotlin/net/thechance/mena/identity/presentation/screen/profile/components/share/utils/ShareSheet.ios.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.presentation.screen.profile.components.dialog.share
+package net.thechance.mena.identity.presentation.screen.profile.components.share.utils
 
 import androidx.compose.runtime.Composable
 import kotlinx.cinterop.BetaInteropApi
